requirements.txt
Flask==1.1.1
opencv_contrib_python==4.1.1.26
Click==7.0
pytesseract==0.3.0
numpy==1.17.2
pyzbar==0.1.8
opencv_python_headless==4.1.1.26
scipy==1.3.1
passlib==1.7.1
Flask==1.1.1
imutils==0.5.3
face_recognition==1.2.3
PyMySQL==0.9.3
dlib==19.18.0
Pillow==6.2.1
scikit_learn==0.21.3
passlib==1.7.1
PyMySQL==0.9.3
